contributors/rodneyrehm/incoming/css3-transitions/properties-value-implicit-001.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8">
    	<meta http-equiv="content-type" content="text/html; charset=utf-8">
        <title>CSS Test: font-size-relative properties transition by implicit value change</title>
        <meta name="assert" content="Test checks that font-size-relative properties (all em-lengths) run a transition when font-size is changed">
        <link rel="help" href="http://www.w3.org/TR/css3-transitions/#transitions">
        <link rel="author" title="Rodney Rehm" href="http://rodneyrehm.de/en/">
        <meta name="flags" content="dom invalid">
        
        <script src="/resources/testharness.js" type="text/javascript"></script>
        <script src="/resources/testharnessreport.js" type="text/javascript"></script>
        
        <script src="./support/helper.js" type="text/javascript"></script>
        <script src="./support/runAsyncHarness.js" type="text/javascript"></script>
        <script src="./support/generalTest.js" type="text/javascript"></script>
        <script src="./support/properties.js" type="text/javascript"></script>
    </head>
    <body>
        <!-- required by testharnessreport.js -->
        <div id="log"></div>
        <!-- elements used for testing -->
        <div id="fixture">
            <div id="container">
                <div id="transition">Hello World</div>
            </div>
        </div>

        <script>
            // see README.md for an explanation of how this test suite works
            
            // this test takes its time, give it 5 minutes to run
            setup({timeout: 300000});
            // init dom
            domFixture('#fixture');

            var tests = getFontSizeRelativePropertyTests();
            // for testing, limit to a couple of iterations
            // tests = tests.slice(10, 30);
            
            // missing support for requestAnimationFrame forces us to
            // run animations at a slower pace, so setTimeout has a
            // chance of being invoked before TransitionEnd
            var duration = getRequestAnimationFrame() ? '0.1s' : '0.5s';

            runAsyncHarness({
                // array of test data
                tests: tests,
                // milliseconds to wait before calling teardown and ending test
                duration: parseFloat(duration) * 1000,
                // prepare individual test
                setup: generalTest.setup,
                // cleanup after individual test
                teardown: generalTest.teardown,
                // run actual test, assertions can be used here!
                test: function(data, options) {
                    // as we're dealing with em-based lengths, we need to define a font-size
                    data.from['font-size'] = '20px';
                    data.to['font-size'] = '30px';
                    // remove property from target style so it won't transition on its own
                    delete data.to[data.property];
                    
                    var styles = {
                        '#fixture': {},

                        '#container': data.parentStyle,
                        '#container.to': {},
                        '#container.transition': {},
                        
                        '#transition': data.from,
                        '#transition.to' : data.to,
                        // only transition the em-based property, not font-size!
                        '#transition.transition' : {transition: data.property + ' ' + duration + ' linear 0s'}
                    };
                    
                    // prepare styles
                    generalTest.setStyle(data, styles);
                    // make sure values differ, if they don't, the property could most likely not be parsed
                    assert_not_equals(data.transition.from, data.transition.to, "initial and target values may not match");
                    // collect computed values during transition
                    generalTest.startValueCollection(data);
                    // kick off the transition
                    generalTest.startTransition(data);
                    
                    // make sure we didn't get the target value immediately.
                    // If we did, there wouldn't be a transition!
                    var current = data.transition.computedStyle(data.property);
                    assert_not_equals(current, data.transition.to, "must not be target value after start");
                },
                // run tests immediately before teardown, assertions can be used here!
                concluding: function(data) {
                    // make sure there were no events on parent
                    data.test.step(generalTest.assertExpectedEventsFunc(data, 'container', ""));
                    // make sure we got the event for the tested property only
                    data.test.step(generalTest.assertExpectedEventsFunc(data, 'transition', data.property + ":" + duration));
                    // make sure the property's value were neither initial nor target while transitioning
                    data.test.step(generalTest.assertIntermediateValuesFunc(data, 'transition'));
                },
                
                // called once all tests are done
                done: generalTest.done
            });
        </script>
    </body>
</html>
